Estimate Rt via exponential smoothing

This option estimates the effective reproduction number over time using exponential smoothing. It implements Holt's linear trend method with dampening through an innovations state space model with a level, trend, and dampening component.

Usage

R_estimate_ets(
  level_prior_mu = 1,
  level_prior_sigma = 0.8,
  trend_prior_mu = 0,
  trend_prior_sigma = 0.1,
  sd_prior_mu = 0,
  sd_prior_sigma = 0.1,
  sd_changepoint_dist = 7 * 26,
  sd_changepoint_sd = 0.025,
  link = "inv_softplus",
  R_max = 6,
  smooth_prior_mu = 0.5,
  smooth_prior_sigma = 0.05,
  trend_smooth_prior_mu = 0.5,
  trend_smooth_prior_sigma = 0.05,
  dampen_prior_mu = 0.9,
  dampen_prior_sigma = 0,
  differenced = FALSE,
  noncentered = TRUE,
  modeldata = modeldata_init()
)

Arguments

level_prior_mu

Prior (mean) on the initial level of Rt.

level_prior_sigma

Prior (standard deviation) on the initial level of Rt.

trend_prior_mu

Prior (mean) on the initial trend of Rt.

trend_prior_sigma

Prior (standard deviation) on the initial trend of Rt.

sd_prior_mu

Prior (mean) on the standard deviation of the innovations.

sd_prior_sigma

Prior (standard deviation) on the standard deviation of the innovations. Please note that for consistency the overall prior on the standard deviation of innovations will have a standard deviation of sd_prior_sigma + sd_changepoint_sd even if no changepoints are modeled (see below).

sd_changepoint_dist

The variability of Rt can change over time, e.g. during the height of an epidemic wave, countermeasures may lead to much faster changes in Rt than observable at other times. This potential variability is accounted for using change points placed at regular intervals. The standard deviation of the state space model innovations then evolves linearly between the change points. The default change point distance is 26 weeks (182 days). Short changepoint distances (e.g. 4 weeks or less) must be chosen with care, as they can make the Rt time series too flexible. If set to zero, no change points are modeled.

sd_changepoint_sd

This parameter controls the variability of the change points. When change points are modeled, EpiSewer will estimate a baseline standard deviation (see sd_prior_mu and sd_prior_sigma), and model change point values as independently distributed with mean equal to this baseline and standard deviation sd_changepoint_sd.

link

Link function. Currently supported are inv_softplus (default) and scaled_logit. Both of these links are configured to behave approximately like the identity function around R=1, but become increasingly non-linear below (and in the case of scaled_logit also above) R=1.

R_max

If link=scaled_logit is used, a maximum reproduction number must be assumed. This should be higher than any realistic R value for the modeled pathogen. Default is 6.

smooth_prior_mu

Prior (mean) on the smoothing parameter. Must be between 0 and 1.

smooth_prior_sigma

Prior (standard deviation) on the smoothing parameter. If this is set to zero, the smoothing parameter will be fixed to smooth_prior_mu and not estimated. If positive, a beta prior with the corresponding mean and standard deviation is used.

trend_smooth_prior_mu

Prior (mean) on the trend smoothing parameter. Must be between 0 and 1.

trend_smooth_prior_sigma

Prior (standard deviation) on the trend smoothing parameter. If this is set to zero, the trend smoothing parameter will be fixed to trend_smooth_prior_mu and not estimated. If positive, a beta prior with the corresponding mean and standard deviation is used.

dampen_prior_mu

Prior (mean) on the dampening parameter. Must be between 0 and 1.

dampen_prior_sigma

Prior (standard deviation) on the dampening parameter. If this is set to zero, the dampening parameter will be fixed to dampen_prior_mu and not estimated. If positive, a beta prior with the corresponding mean and standard deviation is used.

differenced

If FALSE (default), exponential smoothing is applied to the absolute Rt time series. If TRUE, it is instead applied to the differenced time series. This makes the level become the trend, and the trend become the curvature.

noncentered

If TRUE (default), a non-centered parameterization is used to model the innovations in the state space process (for better sampling efficiency).

modeldata

A modeldata object to which the above model specifications should be added. Default is an empty model given by modeldata_init(). Can also be an already partly specified model returned by other EpiSewer modeling functions.

Value

A modeldata object containing data and specifications of the model to be fitted. Can be passed on to other EpiSewer modeling functions to add further data and model specifications.

The modeldata object also includes information about parameter initialization (init), meta data (.metainfo), and checks to be performed before model fitting (.checks).

Details

The innovations state space model consists of three components: a level, a trend, and a dampening component.

The level is smoothed based on the levels from earlier time steps, with exponentially decaying weights, as controlled by a smoothing parameter (often called alpha). Note that smaller values of alpha indicate stronger smoothing. In particular, alpha = 1 means that only the last level is used.

The trend is smoothed based on the trends from earlier time steps, with exponentially decaying weights, as controlled by a trend smoothing parameter (often called beta). Note that smaller values of beta indicate stronger smoothing. In particular, beta = 1 means that only the last trend is used.

The dampening determines how long a previous trend continues into the future before it levels of to a stationary time series. The strength of dampening is controlled by a dampening parameter (often called phi). Note that smaller values of phi indicate stronger dampening. In particular, phi = 1 means no dampening. Values below phi = 0.8 are seldom in practice as the dampening becomes very strong.

Often, alpha, beta, and phi are jointly unidentifiable. It may therefore be necessary to fix at least one of the parameters (typically phi) or supply strong priors.

Note that the smoothness of retrospective Rt estimates is often more influenced by the prior on the standard deviation of innovations than the smoothing and trend smoothing parameters. The smoothing parameters mostly have an influence on estimates close to the present / date of estimation, when limited data signal is available. Here, the standard deviation of the innovations influences how uncertain Rt estimates are close to the present.

The priors of this component have the following functional form:

initial level of Rt: Normal

initial trend of Rt: Normal

standard deviation of innovations: Truncated normal

smoothing parameter: Beta

trend smoothing parameter: Beta

dampening parameter: Beta

See also