Forecasting module

This module function is used to specify the components of the forecast module in EpiSewer.

Each component can be specified using one or several helper functions (see available options below). See the documentation of the individual helper functions to adjust various settings.

Usage

model_forecast(horizon = horizon_none())

Arguments

horizon

The forecast horizon. How many days into the future should EpiSewer forecast? Note that this functionality is intended for short-term forecasts. Projections over longer horizons can be highly inaccurate. Available options: - horizon_none()

horizon_assume()

Value

A modeldata object containing the data and specifications of the forecast module.

Details

Forecasts account for the estimated variation of transmission dynamics over time and therefore become more uncertain at longer forecast horizons. However, it is important to keep in mind that in expectation the model will project the current transmission dynamics to continue unchanged (with the exception of R_estimate_ets, which can project the current linear trend of Rt). This assumption can be violated by various factors such as depletion of susceptible individuals, changes in behavior, or public health interventions.

See also