Model the infection process

This module function is used to specify the components of the infections module in EpiSewer.

Each component can be specified using one or several helper functions (see available options below). See the documentation of the individual helper functions to adjust model priors and further settings.

Usage

model_infections(
  generation_dist = generation_dist_assume(),
  R = R_estimate_splines(),
  seeding = seeding_estimate_rw(),
  infection_noise = infection_noise_estimate()
)

Arguments

generation_dist

Generation time distribution. The intrinsic distribution of the time between infection of a primary case and infection of its secondary cases. Modeling options:

generation_dist_assume()

R

Effective reproduction number over time. This is the main parameter of interest estimated by EpiSewer. R is smoothed using a time series smoothing prior. Currently supported are: random walk (rw), exponential smoothing (ets), and smoothing splines. Modeling options:

R_estimate_approx()

R_estimate_rw()

R_estimate_splines()

R_estimate_ets()

seeding

Seeding of initial infections. The renewal model used by EpiSewer requires a seeding phase of the length of the maximum generation time. For these initial infections, a simple seeding model instead of the renewal model must be used. Modeling options:

seeding_estimate_constant()

seeding_estimate_rw()

infection_noise

Noise in the infection process. EpiSewer implements a stochastic infection model, i.e. allows for variation in the number of new infections generated at each time step. This accounts for stochastic uncertainty in the infection process and often speeds up model fitting. Modeling options:

infection_noise_none()

infection_noise_estimate()

Value

A modeldata object containing the data and specifications of the infections module.

See also