Model the measurement process

This module function is used to specify the components of the measurements module in EpiSewer.

Each component can be specified using one or several helper functions (see available options below). See the documentation of the individual helper functions to adjust model priors and further settings.

Usage

model_measurements(
  concentrations = concentrations_observe(),
  noise = noise_estimate(),
  LOD = LOD_none()
)

Arguments

concentrations

Concentration measurements from wastewater samples. Modeling options:

concentrations_observe()

noise

Measurement noise due to unexplained variation in sampling and lab analysis. Modeling options:

noise_estimate()

noise_estimate_dPCR()

noise_estimate_constant_var()

LOD

Limit of detection. Concentrations below a certain threshold may not be detectable and thus erroneously measured as 0. EpiSewer can adjust for the limit of detection using a hurdle model. Modeling options:

LOD_none()

LOD_assume()

LOD_estimate_dPCR()

Value

A modeldata object containing the data and specifications of the measurements module.

See also