Estimate Rt via a random walk

This option estimates the effective reproduction number over time using a random walk.

Usage

R_estimate_rw(
  intercept_prior_mu = 1,
  intercept_prior_sigma = 0.8,
  sd_prior_mu = 0,
  sd_prior_sigma = 0.1,
  sd_changepoint_dist = 7 * 26,
  sd_changepoint_sd = 0.025,
  link = "inv_softplus",
  R_max = 6,
  differenced = FALSE,
  noncentered = TRUE,
  modeldata = modeldata_init()
)

Arguments

intercept_prior_mu

Prior (mean) on the intercept of the random walk.

intercept_prior_sigma

Prior (standard deviation) on the intercept of the random walk.

sd_prior_mu

Prior (mean) on the standard deviation of the random walk.

sd_prior_sigma

Prior (standard deviation) on the standard deviation of the random walk.

sd_changepoint_dist

The variability of Rt can change over time, e.g. during the height of an epidemic wave, countermeasures may lead to much faster changes in Rt than observable at other times. This potential variability is accounted for using change points placed at regular intervals. The standard deviation of the random walk then evolves linearly between the change points. The default change point distance is 26 weeks (182 days). Short changepoint distances (e.g. 4 weeks or less) must be chosen with care, as they can make the Rt time series too flexible. If set to zero, no change points are modeled.

sd_changepoint_sd

This parameter controls the variability of the change points. When change points are modeled, EpiSewer will estimate a baseline standard deviation (see sd_prior_mu and sd_prior_sigma), and model change point values as independently distributed with mean equal to this baseline and standard deviation sd_changepoint_sd.

link

Link function. Currently supported are inv_softplus (default) and scaled_logit. Both of these links are configured to behave approximately like the identity function around R=1, but become increasingly non-linear below (and in the case of scaled_logit also above) R=1.

R_max

If link=scaled_logit is used, a maximum reproduction number must be assumed. This should be higher than any realistic R value for the modeled pathogen. Default is 6.

differenced

If FALSE (default), the random walk is applied to the absolute Rt time series. If TRUE, it is instead applied to the differenced time series, i.e. now the trend is modeled as a random walk.

noncentered

If TRUE (default), a non-centered parameterization is used to model the innovations of the random walk (for better sampling efficiency).

modeldata

A modeldata object to which the above model specifications should be added. Default is an empty model given by modeldata_init(). Can also be an already partly specified model returned by other EpiSewer modeling functions.

Value

A modeldata object containing data and specifications of the model to be fitted. Can be passed on to other EpiSewer modeling functions to add further data and model specifications.

The modeldata object also includes information about parameter initialization (init), meta data (.metainfo), and checks to be performed before model fitting (.checks).

Details

The smoothness of Rt estimates is influenced by the prior on the standard deviation of the random walk. It also influences the uncertainty of Rt estimates towards the present / date of estimation, when limited data signal is available. The prior on the intercept of the random walk should reflect your expectation of Rt at the beginning of the time series. If estimating from the start of an epidemic, you might want to use a prior with mean > 1 for the intercept.

The priors of this component have the following functional form:

intercept of the random walk: Normal

standard deviation of the random walk: Truncated normal

See also